folder creation complete, moved login and out btns to buttons folder, moved auth js files to auth foler. Edited imports to their appropriate locations
