--- sample_data/.gitkeep
@@ -1,1 +1,0 @@
-add your sample data in this folder and delete this file
