+++ sample_data/orders.json
@@ -1,0 +1,110 @@
+{
+  "-MjX7e1Y-jXcJzlgEKHO" : {
+    "customerEmail" : "orayson0@jugem.jp",
+    "customerName" : "Orville Rayson",
+    "customerPhoneNumber" : "246-760-2107",
+    "orderId" : "-MjX7e1Y-jXcJzlgEKHO",
+    "orderStatus" : true,
+    "orderTotal" : "$34.08",
+    "orderType" : "in Person",
+    "paymentMethod" : "5048372271048163",
+    "timeStamp" : "2020-10-24 03:54:26",
+    "tipTotal" : "$11.13"
+  },
+  "-MjX7plHF4zfL0Viz6-E" : {
+    "customerEmail" : "sdaoust1@pinterest.com",
+    "customerName" : "Sarina Daoust",
+    "customerPhoneNumber" : "620-347-3872",
+    "orderId" : "-MjX7plHF4zfL0Viz6-E",
+    "orderStatus" : false,
+    "orderTotal" : "$50.66",
+    "orderType" : "phone",
+    "paymentMethod" : "5048373025510789",
+    "timeStamp" : "2021-02-16 06:20:21",
+    "tipTotal" : "$9.55"
+  },
+  "-MjX7yZmfaDN0qQyJe-4" : {
+    "customerEmail" : "afilippazzo2@over-blog.com",
+    "customerName" : "Alma Filippazzo",
+    "customerPhoneNumber" : "398-543-6755",
+    "orderId" : "-MjX7yZmfaDN0qQyJe-4",
+    "orderStatus" : false,
+    "orderTotal" : "$46.16",
+    "orderType" : "phone",
+    "paymentMethod" : "5108759064027940",
+    "timeStamp" : "2021-07-26 13:39:34",
+    "tipTotal" : "$23.40"
+  },
+  "-MjX838zRCKaW0EvEtID" : {
+    "customerEmail" : "vanstead3@tuttocitta.it",
+    "customerName" : "Valene Anstead",
+    "customerPhoneNumber" : "241-414-5229",
+    "orderId" : "-MjX838zRCKaW0EvEtID",
+    "orderStatus" : false,
+    "orderTotal" : "$15.94",
+    "orderType" : "in Person",
+    "paymentMethod" : "5048370841441975",
+    "timeStamp" : "2021-06-28 16:27:02",
+    "tipTotal" : "$7.52"
+  },
+  "-MjX8A7b7X2H-T10WAiI" : {
+    "customerEmail" : "csouthcomb4@ca.gov",
+    "customerName" : "Cherey Southcomb",
+    "customerPhoneNumber" : "691-868-5848",
+    "orderId" : "-MjX8A7b7X2H-T10WAiI",
+    "orderStatus" : false,
+    "orderTotal" : "$21.13",
+    "orderType" : "phone",
+    "paymentMethod" : "5048379838759687",
+    "timeStamp" : "2020-12-10 06:36:42",
+    "tipTotal" : "$7.91"
+  },
+  "-MjX8E5AgtJFWLjN3qjR" : {
+    "customerEmail" : "bqueen5@123-reg.co.uk",
+    "customerName" : "Benny Queen",
+    "customerPhoneNumber" : "184-911-6678",
+    "orderId" : "-MjX8E5AgtJFWLjN3qjR",
+    "orderStatus" : false,
+    "orderTotal" : "$19.88",
+    "orderType" : "in Person",
+    "paymentMethod" : "5048378969869042",
+    "timeStamp" : "2020-11-10 08:54:59",
+    "tipTotal" : "$37.70"
+  },
+  "-MjX8Hn--yv95IbA15gN" : {
+    "customerEmail" : "mclaye6@timesonline.co.uk",
+    "customerName" : "Morris Claye",
+    "customerPhoneNumber" : "415-798-7043",
+    "orderId" : "-MjX8Hn--yv95IbA15gN",
+    "orderStatus" : true,
+    "orderTotal" : "$72.92",
+    "orderType" : "in Person",
+    "paymentMethod" : "5108755629029298",
+    "timeStamp" : "2021-07-19 02:07:45",
+    "tipTotal" : "$20.80"
+  },
+  "-MjX8LHGVFJXJhYT1YlX" : {
+    "customerEmail" : "aluckes7@dailymotion.com",
+    "customerName" : "Antonius Luckes",
+    "customerPhoneNumber" : "745-652-1274",
+    "orderId" : "-MjX8LHGVFJXJhYT1YlX",
+    "orderStatus" : false,
+    "orderTotal" : "$99.46",
+    "orderType" : "in Person",
+    "paymentMethod" : "5048373074108931",
+    "timeStamp" : "2021-08-02 01:12:22",
+    "tipTotal" : "$85.77"
+  },
+  "-MjX8Sd7EKeeQ1sWFHEe" : {
+    "customerEmail" : "rshulver8@prnewswire.com",
+    "customerName" : "Renell Shulver",
+    "customerPhoneNumber" : "650-504-2598",
+    "orderId" : "-MjX8Sd7EKeeQ1sWFHEe",
+    "orderStatus" : true,
+    "orderTotal" : "$40.49",
+    "orderType" : "in Person",
+    "paymentMethod" : "5048375028819539",
+    "timeStamp" : "2021-02-14 10:02:58",
+    "tipTotal" : "$4.79"
+  }
+}
