+++ src/scripts/components/buttons/logoutButton.js
