+++ src/scripts/helpers/auth/signIn.js
