+++ src/scripts/helpers/auth/signOut.js
